+++ .gitignore
@@ -59,3 +59,4 @@ Temporary Items
 
 /index.js
 dist
+/build/*.js
