+++ .npmignore
@@ -4,3 +4,4 @@ index.html
 !dist/*
 screenshot
 index-dev.js
+!/build/*
