+++ package.json
@@ -2,10 +2,10 @@
   "name": "online-explorer",
   "version": "1.0.9",
   "description": "An online file explorer",
-  "main": "index.js",
+  "main": "build/explorer.js",
   "author": "helloqiu <helloqiu95@gmail.com>",
   "scripts": {
-    "build": "babel index-dev.js --out-file index.js && chmod +x index.js"
+    "build": "babel index-dev.js --out-file index.js && chmod +x index.js && babel src/explorer.js --out-file build/explorer.js"
   },
   "repository": {
     "type": "git",
@@ -34,7 +34,11 @@
   },
   "devDependencies": {
     "babel-cli": "^6.22.2",
-    "babel-preset-latest": "^6.22.0"
+    "babel-preset-latest": "^6.22.0",
+    "eslint": "^3.15.0",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-promise": "^3.4.1",
+    "eslint-plugin-standard": "^2.0.1"
   },
   "babel": {
     "presets": [
